refactor(JavaScript): move UserStory into its own module
